prep for pulumi mgmt
