+++ examples/examples_dotnet_test.go
@@ -1,0 +1,21 @@
+// Copyright 2016-2017, Pulumi Corporation.  All rights reserved.
+// +build dotnet all
+
+package examples
+
+import (
+	"testing"
+
+	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
+)
+
+func getCsharpBaseOptions(t *testing.T) integration.ProgramTestOptions {
+	base := getBaseOptions()
+	baseCsharp := base.With(integration.ProgramTestOptions{
+		Dependencies: []string{
+			"Pulumi.Tls",
+		},
+	})
+
+	return baseCsharp
+}
